+++ CHANGELOG.md
@@ -4,8 +4,19 @@
 
 ### Added
 
-- Add simple unit test.
-- New parameter `name` for `BoolArgument`.
+- Add unit test for parser.
+- Add unit test for usage.
+- New field `name` for `BoolArgument`.
+- New field `name` for `NumberArgument`.
+- Add `SingleUsage` trait.
+- Add `MultipleUsage` trait.
+- Add `IntoMultipleUsage` trait.
+- Add `ChildUsage` trait.
+- Add `Prefix` type and `prefix` fn.
+- Add `Chain` type (from `chain` fn on `MultipleUsage`).
+- Implement Usage traits for `BoolArgument`, `LiteralArgument` and `NumberArgument`.
+- Implement Usage traits for `CommandThen`, `ThenWrapper` and `LiteralThen`.
+- Implement Usage traits for `DefaultExecutor`, `LiteralExecutor`, `LiteralThenExecutor` and `ThenExecutor`.
 
 ### Changed